Split SMB client and server
